+++ examples/flutter_app/lib/example_configurator.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_scene/scene.dart' hide BoxShape;
+import 'package:flutter_scene/scene.dart';
 import 'package:vector_math/vector_math.dart' as vm;
 
 import 'environment_menu.dart';