+++ packages/flutter_scene_rapier/test/body_simulation_test.dart
@@ -11,6 +11,7 @@
 // context).
 
 import 'package:flutter_scene/scene.dart';
+import 'package:flutter_scene/physics.dart';
 import 'package:flutter_scene_rapier/flutter_scene_rapier.dart';
 import 'package:test/test.dart';
 import 'package:vector_math/vector_math.dart';